perf(format): remove dead code
